Plumb through EarlyEOF to callers
If we receive an EarlyEOF, send this through to the calling program,
so we can hit its callback and terminate it, instead of hanging
forever.

Verified this patch works by adding it to the Sourcegraph node_modules
directory, and sending Ctrl+C to the server - it terminates the
running process with an error instead of hanging forever.